requirements.txt
asgiref==3.8.1
Django==5.0.6
djangorestframework==3.15.1
pytz==2023.3
requests==2.31.0
sqlparse==0.5.0
typing_extensions==4.11.0
